requirements.txt
absl-py==0.9.0
argcomplete==1.10.0
asgiref==3.2.7
astor==0.8.1
audioread==2.1.8
beautifulsoup4==4.8.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
Django==1.8.6
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
gast==0.2.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.9
imageio==2.8.0
IMAPClient==2.1.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.31.0
lxml==4.5.0
Markdown==3.2.1
matplotlib==3.2.1
networkx==2.4
nltk==3.5
numba==0.48.0
numpy==1.18.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.2.0
pandas==1.0.3
pdfminer.six==20181108
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
python-vlc==3.0.7110
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textract==1.6.3
tqdm==4.45.0
tzlocal==1.5.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.8
